+++ javascript/src/defaultFormatCode.ts
@@ -1,0 +1,5 @@
+import { Snippet } from '@cucumber/messages'
+
+import { FormatCodeFunction } from './types'
+
+export const defaultFormatCode: FormatCodeFunction = (snippet: Snippet) => snippet.code