+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [JavaScript] Support custom code formatting ([#92](https://github.com/cucumber/pretty-formatter/pull/92))
+
 ### Fixed
 - Avoid stack trace with pending and skipped ([#91](https://github.com/cucumber/pretty-formatter/pull/91))